requirements.txt
bitsandbytes==0.49.1
datasets==4.4.2
einops==0.4.1
fire==0.5.0
gradio==3.34.0
openai==0.27.0
peft==0.18.0
pillow==9.3.0
retry==0.9.2
sentencepiece==0.1.99
protobuf==3.20.3
transformers==4.57.3
triton==3.5.1
tqdm==4.67.1
wandb>=0.16.0
click==8.1.6
accelerate==1.12.0
huggingface_hub>=0.23.0
numpy<2.0
torchvision==0.24.1